Finish lab2
