+++ lab2/kernel/include/dtb.h
@@ -1,21 +1,30 @@
-#ifndef _DTB_H_
-#define _DTB_H_
-
 #define uint32_t unsigned int
 
-// manipulate device tree with dtb file format
-// linux kernel fdt.h
 #define FDT_BEGIN_NODE 0x00000001
 #define FDT_END_NODE 0x00000002
 #define FDT_PROP 0x00000003
 #define FDT_NOP 0x00000004
 #define FDT_END 0x00000009
 
+//stored as big endian
+struct fdt_header {
+    uint32_t magic;
+    uint32_t totalsize;
+    uint32_t off_dt_struct;
+    uint32_t off_dt_strings;
+    uint32_t off_mem_rsvmap;
+    uint32_t version;
+    uint32_t last_comp_version;
+    uint32_t boot_cpuid_phys;
+    uint32_t size_dt_strings;
+    uint32_t size_dt_struct;
+};
+
+typedef struct fdt_header fdt_header;
+
 typedef void (*dtb_callback)(uint32_t node_type, char *name, void *value, uint32_t name_size);
 
 uint32_t uint32_endian_big2lttle(uint32_t data);
-void traverse_device_tree(void *base, dtb_callback callback);  //traverse dtb tree
-void dtb_callback_show_tree(uint32_t node_type, char *name, void *value, uint32_t name_size);
-void dtb_callback_initramfs(uint32_t node_type, char *name, void *value, uint32_t name_size);
-
-#endif
+void fdt_traverse(void *base, dtb_callback callback);  //traverse dtb tree
+void show_tree_callback(uint32_t node_type, char *name, void *value, uint32_t name_size);
+void initramfs_callback(uint32_t node_type, char *name, void *value, uint32_t name_size);
