+++ lab2/kernel/include/string.h
@@ -1,0 +1,3 @@
+char* strcpy(char* dest, const char* src);
+int strcmp(const char* str1, const char* str2);
+unsigned long long strlen(const char *str);
